Highlight request on run() and last() (#60)
New feature "highlight request on run() and last()". In order to implement this, we had to switch from synchronous call to plenary-curl to asynchronous with callback. Implementation is similar to the build-in vim.highlight.on_yank. This feature is enabled by default.

Additionally, the old feature of moving the cursor to the current request line is disabled by default as the highlighting shows more accurate what is send and in my opinion it is nicer to have the cursor stay in place.