fix: ensure help window width is an int
Problem:

The help window causes an error when calling `vim.api.nvim_open_win`
because the width being passed is not an integer.

Neovim's `vim.api.nvim_open_win` function requires that the width and
height be specified in whole integers.

The width is being calculated through a division operation causing it to
potentially be a number that is not an int.

Solution:

Floor the calculation of the help window width so the width is always
rounded down to the nearest integer and can thus be used with
`vim.api.nvim_open_win`.